% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{agostini_example}
\alias{agostini_example}
\title{Datos de ejemplo para calc_agostini()}
\format{
Una lista con los componentes:
\describe{
\item{audiencias}{Vector numerico con las audiencias de cada soporte}
\item{pob_total}{Tamano de la poblacion}
\item{k}{Coeficiente empirico de duplicacion de Agostini}
}
}
\usage{
agostini_example
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_agostini}}, lista para usar con \code{do.call()}.
}
\examples{
data(agostini_example)
do.call(calc_agostini, agostini_example)
}
\keyword{datasets}
