% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_canex.R
\encoding{UTF-8}
\name{calculate_duplication}
\alias{calculate_duplication}
\title{Calcular la duplicacion (correlacion) entre dos vehiculos}
\usage{
calculate_duplication(pij, pi, pj)
}
\arguments{
\item{pij}{Numerico. Probabilidad de exposicion conjunta a ambos vehiculos}

\item{pi}{Numerico. Probabilidad marginal de exposicion al primer vehiculo}

\item{pj}{Numerico. Probabilidad marginal de exposicion al segundo vehiculo}
}
\value{
Numerico. Coeficiente de correlacion (entre -1 y 1); 0 si la
varianza de alguno de los dos vehiculos es nula.
}
\description{
Calcula el coeficiente de correlacion entre las exposiciones
a dos vehiculos a partir de su probabilidad de exposicion conjunta y sus
probabilidades marginales de exposicion.
}
\examples{
corr <- calculate_duplication(0.15, 0.3, 0.4)

}
