% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_canex.R
\encoding{UTF-8}
\name{calculate_marginal_prob}
\alias{calculate_marginal_prob}
\title{Calcular la probabilidad marginal Beta-Binomial}
\usage{
calculate_marginal_prob(x, k, alpha, beta)
}
\arguments{
\item{x}{Entero. Numero de exitos (contactos)}

\item{k}{Entero. Numero de ensayos (inserciones)}

\item{alpha}{Numerico. Parametro alpha de la BBD}

\item{beta}{Numerico. Parametro beta de la BBD}
}
\value{
Numerico. Probabilidad P(X = x)
}
\description{
Calcula, en espacio logaritmico por estabilidad numerica, la
probabilidad de obtener x exitos en k ensayos bajo una distribucion
Beta-Binomial de parametros alpha y beta.
}
\examples{
prob <- calculate_marginal_prob(2, 5, 0.5, 1.2)

}
