% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_canex.R
\encoding{UTF-8}
\name{calculate_mean_variance}
\alias{calculate_mean_variance}
\title{Calcular media y varianza de la Distribucion Beta-Binomial}
\usage{
calculate_mean_variance(k, alpha, beta)
}
\arguments{
\item{k}{Entero. Numero de inserciones (ensayos)}

\item{alpha}{Numerico. Parametro alpha de la BBD}

\item{beta}{Numerico. Parametro beta de la BBD}
}
\value{
Lista con los componentes:
\itemize{
\item mean: Media de la distribucion
\item variance: Varianza de la distribucion
}
}
\description{
Calcula la media y la varianza de una Beta-Binomial dados
el numero de ensayos y sus parametros de forma.
}
\examples{
mv <- calculate_mean_variance(2, 0.5, 1.2)

}
