% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{cpm_example}
\alias{cpm_example}
\title{Datos de ejemplo para calc_cpm()}
\format{
Una lista con los componentes:
\describe{
\item{precios}{Vector numerico con precios de cada insercion}
\item{audiencias}{Vector numerico con audiencias de cada soporte}
}
}
\usage{
cpm_example
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_cpm}}, lista para usar con \code{do.call()}.
}
\examples{
data(cpm_example)
do.call(calc_cpm, cpm_example)
}
\keyword{datasets}
