% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_modelos.R
\encoding{UTF-8}
\name{matriz_a_vector}
\alias{matriz_a_vector}
\title{Convertir una matriz simétrica en un vector (triángulo superior)}
\usage{
matriz_a_vector(matriz)
}
\arguments{
\item{matriz}{Matriz cuadrada simétrica}
}
\value{
Vector numérico con los elementos del triángulo superior de \code{matriz}
}
\description{
Linealiza una matriz simétrica (p.ej. de duplicaciones o de
oportunidades de contacto) recorriendo su triángulo superior, incluyendo
la diagonal, en el orden (1,1), (1,2), ..., (1,n), (2,2), (2,3), ....
Función auxiliar empleada por \code{\link{calc_metheringham}}.
}
\examples{
m <- matrix(c(1, 2, 2, 3), nrow = 2)
matriz_a_vector(m)

}
