% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{metheringham}
\alias{metheringham}
\title{Datos de ejemplo para calc_metheringham()}
\format{
Una lista con los componentes:
\describe{
\item{audiencias}{Vector numerico con las audiencias de cada soporte}
\item{inserciones}{Vector numerico con el numero de inserciones por soporte}
\item{matriz_duplicacion}{Matriz simetrica con la duplicacion entre soportes}
}
}
\usage{
metheringham
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_metheringham}}, lista para usar con \code{do.call()}.
}
\examples{
data(metheringham)
do.call(calc_metheringham, metheringham)
}
\keyword{datasets}
