% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{nbd}
\alias{nbd}
\title{Historical NBD example data name}
\format{
Una lista con los componentes:
\describe{
\item{audiencias}{Vector numerico con las audiencias de cada soporte}
\item{inserciones}{Vector numerico con el numero de inserciones por soporte}
\item{pob_total}{Tamano de la poblacion}
\item{k}{Parametro de heterogeneidad (forma de la distribucion Gamma)}
}
}
\usage{
nbd
}
\description{
Legacy name for \code{nbd_example}. New statistical work
should normally use \code{\link{fit_nbd_exposure}} with observed counts.
}
\examples{
data(nbd)
do.call(calc_nbd, nbd)
}
\keyword{datasets}
