% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grp_metricas.R
\encoding{UTF-8}
\name{plot_grp_metricas}
\alias{plot_grp_metricas}
\title{Visualizacion de GRPs y metricas relacionadas por soporte}
\usage{
plot_grp_metricas(
  audiencias,
  inserciones,
  precios,
  nombres,
  pob_total,
  titulo = "Comparacion de Soportes Publicitarios"
)
}
\arguments{
\item{audiencias}{Vector numerico con las audiencias de cada soporte}

\item{inserciones}{Vector numerico del numero de inserciones por soporte}

\item{precios}{Vector numerico con el precio por insercion de cada soporte}

\item{nombres}{Character vector con los nombres de los soportes}

\item{pob_total}{Tamano de la poblacion objetivo}

\item{titulo}{Character. Titulo del grafico (opcional)}
}
\value{
Un objeto ggplot2 que representa el grafico de burbujas
}
\description{
Genera un grafico de puntos para comparar soportes publicitarios
segun GRPs y coste total, contactos y coste/GRP. El grafico muestra la relacion entre el coste por GRP,
los contactos totales y el coste total de cada soporte, utilizando un sistema de
burbujas con colores distintivos para cada soporte.
}
\examples{
# Ejemplo basico con tres soportes
plot_grp_metricas(
  audiencias = c(300000, 400000, 200000),
  inserciones = c(3, 2, 4),
  precios = c(1000, 1500, 800),
  nombres = c("Marca", "As", "20 Minutos"),
  pob_total = 1000000,
  titulo = "Analisis de Soportes Deportivos"
)

}
