% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bbd_to_reach.R
\name{print.bbd_reach_fit}
\alias{print.bbd_reach_fit}
\title{Print a Beta-Binomial reach fit}
\usage{
\method{print}{bbd_reach_fit}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bbd_reach_fit}.}

\item{...}{Argumentos adicionales pasados a print}
}
\description{
Prints the results of a Beta-Binomial distribution fitted to
an external reach estimate.
}
