% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{Print methods for mediaplanr objects}
\arguments{
\item{x}{The object to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print methods for mediaplanr objects
}
