% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_converg_explorer.R
\encoding{UTF-8}
\name{run_reach_converg_explorer}
\alias{run_reach_converg_explorer}
\title{Explorador de Convergencia de la Cobertura}
\usage{
run_reach_converg_explorer()
}
\value{
Lanza una aplicacion Shiny interactiva (efecto secundario); no
devuelve ningun valor util para el flujo de un script.
}
\description{
Aplicacion Shiny para el analisis de la convergencia de la cobertura.
}
\details{
La aplicacion permite:
\itemize{
\item Configurar un plan de medios aplicando el modelo Beta-Binomial
\item Analizar la evolucion de la cobertura acumulada e incremental
\item Analizar la distribucion de contactos (y acumulada)
\item Visualizar distribuciones mediante graficos de lineas
\item Calcular estadisticas relevantes de la audiencia
}
}
\section{Parametros de Configuracion}{

\itemize{
\item Tamano de poblacion
\item Parametros de forma de la distribucion Beta-Binomial
\item Maximo numero de contactos a mostrar
\item Umbral de convergencia
}
}

\examples{
if (interactive()) {
  run_reach_converg_explorer()
}

}
