% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_canex.R
\encoding{UTF-8}
\name{transform_duplications}
\alias{transform_duplications}
\title{Transformar la matriz de duplicaciones en matriz de correlaciones}
\usage{
transform_duplications(duplications, vehicles_data)
}
\arguments{
\item{duplications}{Matriz cuadrada. Duplicaciones brutas entre vehiculos (0-1)}

\item{vehicles_data}{Data frame con, al menos, la columna R1 de cada vehiculo}
}
\value{
Matriz de coeficientes de correlacion (diagonal = 1)
}
\description{
Convierte la matriz de duplicaciones brutas (proporcion de
poblacion expuesta simultaneamente a cada par de vehiculos) en la matriz
de coeficientes de correlacion que el modelo CANEX usa como termino de
ajuste de la probabilidad conjunta.
}
\examples{
dup_matrix <- matrix(c(1, 0.0157, 0.0157, 1), nrow = 2)
vehicles <- data.frame(R1 = c(0.4902, 0.033))
correlations <- transform_duplications(dup_matrix, vehicles)

}
