% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{agostini_example}
\alias{agostini_example}
\title{Example inputs for calc_agostini()}
\format{
A list with the components:
\describe{
\item{audiences}{Numeric vector with the audience of each vehicle}
\item{population}{Population size}
\item{k}{Agostini's empirical duplication coefficient}
}
}
\usage{
agostini_example
}
\description{
List of arguments for \code{\link{calc_agostini}}, ready for
\code{do.call()}.
}
\examples{
data(agostini_example)
do.call(calc_agostini, agostini_example)
}
\keyword{datasets}
