% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{beta_binomial_example}
\alias{beta_binomial_example}
\title{Example inputs for calc_beta_binomial()}
\format{
A list with the components:
\describe{
\item{A1}{Vehicle audience after the first insertion}
\item{A2}{Vehicle audience after the second insertion}
\item{P}{Total population size}
\item{n}{Total number of planned insertions}
}
}
\usage{
beta_binomial_example
}
\description{
List of arguments for \code{\link{calc_beta_binomial}}, ready
for \code{do.call()}.
}
\examples{
data(beta_binomial_example)
do.call(calc_beta_binomial, beta_binomial_example)
}
\keyword{datasets}
