% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{binomial_plan}
\alias{binomial_plan}
\title{Example inputs for calc_binomial()}
\format{
A list with the components:
\describe{
\item{audiences}{Numeric vector with the audience of each vehicle}
\item{population}{Population size}
}
}
\usage{
binomial_plan
}
\description{
List of arguments for \code{\link{calc_binomial}}, ready for
\code{do.call()}. Called \code{binomial_plan} rather than \code{binomial}
to avoid masking \code{stats::binomial} once the package is loaded.
}
\examples{
data(binomial_plan)
do.call(calc_binomial, binomial_plan)
}
\keyword{datasets}
