% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{canex_example}
\alias{canex_example}
\title{Example inputs for calc_canex()}
\format{
A list with \code{vehicles_data}, \code{duplications}, and
\code{poblacion}.
}
\usage{
canex_example
}
\description{
List of arguments for \code{\link{calc_canex}}, ready for
\code{do.call()}.
}
\examples{
data(canex_example)
do.call(calc_canex, canex_example)
}
\keyword{datasets}
