% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{cpm}
\alias{cpm}
\title{Datos de ejemplo para calc_cpm()}
\format{
Una lista con los componentes:
\describe{
\item{precios}{Vector numérico con precios de cada inserción}
\item{audiencias}{Vector numérico con audiencias de cada soporte}
}
}
\usage{
cpm
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_cpm}}, lista para usar con \code{do.call()}.
}
\examples{
data(cpm)
do.call(calc_cpm, cpm)
}
\keyword{datasets}
