% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{datos_medios}
\alias{datos_medios}
\title{Datos ilustrativos de soportes de prensa}
\format{
Un data frame con 60 filas y 7 variables:
\describe{
\item{soportes}{Caracter. Nombre del soporte}
\item{soportes_.}{Numerico. Coeficiente de soporte (uso interno docente)}
\item{audiencias}{Entero. Audiencia del soporte (personas)}
\item{tarifas}{Numerico. Tarifa de una insercion en el soporte}
\item{indices_utilidad}{Numerico. Indice de utilidad de la audiencia (0-1 aprox.)}
\item{inserciones}{Entero. Numero de inserciones consideradas}
\item{duplicacion}{Entero. Audiencia duplicada estimada frente al resto de soportes (personas)}
}
}
\source{
Elaboracion propia con fines docentes.
}
\usage{
datos_medios
}
\description{
Conjunto de datos ficticio con audiencias, tarifas e indices
de utilidad de 60 soportes de prensa espanola, empleado en los ejemplos
de \code{\link{calc_sainsbury}}, \code{\link{calc_binomial}},
\code{\link{optimize_media_sb}} y \code{\link{calcular_metricas_medios}}.
}
\note{
Los datos son ficticios y se emplean unicamente con fines docentes e
ilustrativos; no representan cifras reales de audiencia o tarifas.
}
\examples{
data(datos_medios)
head(datos_medios)

}
\keyword{datasets}
