% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{grps}
\alias{grps}
\title{Datos de ejemplo para calc_grps()}
\format{
Una lista con los componentes:
\describe{
\item{audiencias}{Vector numérico con las audiencias de cada soporte}
\item{inserciones}{Vector numérico del número de inserciones por soporte}
\item{pob_total}{Tamaño de la población}
}
}
\usage{
grps
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_grps}}, lista para usar con \code{do.call()}.
}
\examples{
data(grps)
do.call(calc_grps, grps)
}
\keyword{datasets}
