% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{hofmans}
\alias{hofmans}
\title{Datos de ejemplo para calc_hofmans()}
\format{
Una lista con los componentes:
\describe{
\item{R1}{Cobertura tras la primera inserción (0-1)}
\item{R2}{Cobertura tras la segunda inserción (0-1)}
\item{N}{Número de inserciones para las que calcular la audiencia acumulada}
}
}
\usage{
hofmans
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_hofmans}}, lista para usar con \code{do.call()}.
}
\examples{
data(hofmans)
do.call(calc_hofmans, hofmans)
}
\keyword{datasets}
