% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{hofmans_accumulation_example}
\alias{hofmans_accumulation_example}
\title{Example inputs for calc_hofmans_accumulation()}
\format{
A list with the components:
\describe{
\item{R1}{Reach after the first insertion (0-1)}
\item{R2}{Reach after the second insertion (0-1)}
\item{N}{Number of insertions for which to calculate cumulative audience}
}
}
\usage{
hofmans_accumulation_example
}
\description{
List of arguments for \code{\link{calc_hofmans_accumulation}},
ready for \code{do.call()}. One vehicle, several insertions -- the
"accumulation" domain. For the other Hofmans model (several vehicles, one
insertion each), see \code{\link{hofmans_duplication_example}}.
}
\examples{
data(hofmans_accumulation_example)
do.call(calc_hofmans_accumulation, hofmans_accumulation_example)
}
\seealso{
\code{\link[=calc_hofmans_accumulation]{calc_hofmans_accumulation()}}, \link{hofmans_duplication_example}
}
\keyword{datasets}
