% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{hofmans_example}
\alias{hofmans_example}
\title{Example inputs for calc_hofmans()}
\format{
A list with the components:
\describe{
\item{R1}{Reach after the first insertion (0-1)}
\item{R2}{Reach after the second insertion (0-1)}
\item{N}{Number of insertions for which to calculate cumulative audience}
}
}
\usage{
hofmans_example
}
\description{
List of arguments for \code{\link{calc_hofmans}}, ready for
\code{do.call()}.
}
\examples{
data(hofmans_example)
do.call(calc_hofmans, hofmans_example)
}
\keyword{datasets}
