% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{MBBD}
\alias{MBBD}
\title{Datos de ejemplo para calc_MBBD()}
\format{
Una lista con los componentes:
\describe{
\item{insertions}{Vector numérico. Número de inserciones para cada soporte}
\item{audiences}{Vector numérico. Audiencia de cada soporte en personas}
\item{RM}{Entero. Estimación de cobertura según Morgensztern en personas}
\item{universe}{Entero. Tamaño del universo objetivo en personas}
\item{A0}{Numérico. Valor inicial del parámetro A}
}
}
\usage{
MBBD
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_MBBD}}, lista para usar con \code{do.call()}.
}
\examples{
data(MBBD)
do.call(calc_MBBD, MBBD)
}
\keyword{datasets}
