% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaPlanR-package.R
\docType{package}
\name{mediaPlanR}
\alias{mediaPlanR}
\alias{mediaPlanR-package}
\title{Herramientas para Planificación de Medios}
\description{
Conjunto de funciones y aplicaciones para planificación de medios.
}
\details{
mediaPlanR: Herramientas para Planificación de Medios
}
\references{
\href{https://github.com/majesus/mediaPlanR}{Repositorio en GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/majesus/mediaPlanR}
  \item Report bugs at \url{https://github.com/majesus/mediaPlanR/issues}
}

}
\author{
Manuel J. Sánchez-Franco \email{majesus@us.es}
}
