% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{metheringham_example}
\alias{metheringham_example}
\title{Example inputs for calc_metheringham()}
\format{
A list with the components:
\describe{
\item{audiences}{Numeric vector with the audience of each vehicle}
\item{insertions}{Numeric vector with the number of insertions per vehicle}
\item{duplication_matrix}{Symmetric matrix with the duplication between vehicles}
\item{population}{Population size}
}
}
\usage{
metheringham_example
}
\description{
List of arguments for \code{\link{calc_metheringham}}, ready
for \code{do.call()}.
}
\examples{
data(metheringham_example)
do.call(calc_metheringham, metheringham_example)
}
\keyword{datasets}
