% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{nbd}
\alias{nbd}
\title{Datos de ejemplo para calc_nbd()}
\format{
Una lista con los componentes:
\describe{
\item{audiencias}{Vector numérico con las audiencias de cada soporte}
\item{inserciones}{Vector numérico con el número de inserciones por soporte}
\item{pob_total}{Tamaño de la población}
\item{k}{Parámetro de heterogeneidad (forma de la distribución Gamma)}
}
}
\usage{
nbd
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_nbd}}, lista para usar con \code{do.call()}.
}
\examples{
data(nbd)
do.call(calc_nbd, nbd)
}
\keyword{datasets}
