% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{sainsbury}
\alias{sainsbury}
\title{Datos de ejemplo para calc_sainsbury()}
\format{
Una lista con los componentes:
\describe{
\item{audiencias}{Vector numérico con las audiencias de cada soporte}
\item{pob_total}{Tamaño de la población}
}
}
\usage{
sainsbury
}
\description{
Lista con los argumentos de ejemplo para
\code{\link{calc_sainsbury}}, lista para usar con \code{do.call()}.
}
\examples{
data(sainsbury)
do.call(calc_sainsbury, sainsbury)
}
\keyword{datasets}
