% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{sainsbury_example}
\alias{sainsbury_example}
\title{Example inputs for calc_sainsbury()}
\format{
A list with the components:
\describe{
\item{audiences}{Numeric vector with the audience of each vehicle}
\item{population}{Population size}
}
}
\usage{
sainsbury_example
}
\description{
List of arguments for \code{\link{calc_sainsbury}}, ready for
\code{do.call()}.
}
\examples{
data(sainsbury_example)
do.call(calc_sainsbury, sainsbury_example)
}
\keyword{datasets}
